Update style.css
Latest version of portfolio website where responsiveness has began but is creating issues. So the navigation bar should go on the side when viewport is smaller, and a toggle button should be there instead that when clicked, opens up the vertical menu and changes the toggle button into an ‘x’, however, in this code, since both icons are in the same class, both icons are doing the same things, whereas one should disappear when the other appears. Plus, the vertical nav block is also in the centre for some reason instead of on the right side as it was before. To fix this issue, I might have to use icons from a different website than currently, this will allow the classes to function smoothly.